# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/inventario/{id}": {
      "get": {
        "tags": [
          "Inventario"
        ],
        "summary": "Obtener movimiento de inventario por id",
        "description": "Retorna un movimiento de inventario por id",
        "operationId": "obtenerMovimientoPorId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id del movimiento de inventario",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Se encontro el movimiento",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventarioResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "No se encontro el movimiento",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Usuario no autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Acceso denegado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Inventario"
        ],
        "summary": "Actualizar movimiento de inventario",
        "description": "Actualiza un movimiento de inventario",
        "operationId": "actualizarMovimiento",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Introducir ID del movimiento que se desea actualizar",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Inventario"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Movimiento actualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventarioResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Datos de entrada inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No se encontro el movimiento",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Usuario no autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Acceso denegado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Inventario"
        ],
        "summary": "Eliminar movimiento de inventario",
        "description": "Elimina un movimiento de inventario",
        "operationId": "eliminarMovimiento",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Introducir ID del movimiento que se desea eliminar",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Movimiento eliminado",
            "content": {
              "application/json": {}
            }
          },
          "404": {
            "description": "No se encontro el movimiento",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Usuario no autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Acceso denegado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/inventario": {
      "get": {
        "tags": [
          "Inventario"
        ],
        "summary": "Listar movimientos de inventario",
        "description": "Retorna una lista de todos los movimientos de inventario",
        "operationId": "listarMovimientosDeInventario",
        "responses": {
          "200": {
            "description": "Se encontraron movimientos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InventarioResponseDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No se encontraron movimientos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Usuario no autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Acceso denegado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Inventario"
        ],
        "summary": "Registrar movimiento de inventario",
        "description": "Registra un nuevo movimiento de inventario",
        "operationId": "registrarMovimiento",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Inventario"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Movimiento registrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventarioResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Datos de entrada inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Usuario no autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Acceso denegado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/ventas/{id}": {
      "get": {
        "tags": [
          "venta-controller"
        ],
        "operationId": "obtenerVentaPorId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Venta"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Rol": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nombre": {
            "type": "string"
          },
          "usuarios": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Usuario"
            },
            "uniqueItems": true
          }
        }
      },
      "Usuario": {
        "type": "object",
        "description": "Introducir objeto de tipo Usuario a guardar",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Rol"
            },
            "uniqueItems": true
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "error": {
            "type": "string"
          },
          "mensaje": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "erroresValidacion": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
      "Categoria": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nombre": {
            "type": "string"
          },
          "productos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Producto"
            }
          }
        }
      },
      "Producto": {
        "type": "object",
        "description": "Introducir objeto de tipo Producto a guardar",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nombre": {
            "type": "string"
          },
          "precio": {
            "type": "number",
            "format": "double"
          },
          "stock": {
            "type": "integer",
            "format": "int32"
          },
          "categoria": {
            "$ref": "#/components/schemas/Categoria"
          }
        }
      },
...
      "Inventario": {
        "type": "object",
        "description": "Introducir objeto de tipo Inventario a guardar",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "producto": {
            "$ref": "#/components/schemas/Producto"
          },
          "cantidad": {
            "type": "integer",
            "format": "int32"
          },
          "tipoMovimiento": {
            "type": "string"
          },
          "fechaMovimiento": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "InventarioResponseDto": {
        "type": "object",
        "properties": {
          "productoId": {
            "type": "integer",
            "format": "int64"
          },
          "cantidad": {
            "type": "integer",
            "format": "int32"
          },
          "tipoMovimiento": {
            "type": "string"
          },
          "fechaMovimiento": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "DetalleVenta": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "venta": {
            "$ref": "#/components/schemas/Venta"
          },
          "producto": {
            "$ref": "#/components/schemas/Producto"
          },
          "cantidad": {
            "type": "integer",
            "format": "int32"
          },
          "precio": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "Venta": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "usuario": {
            "$ref": "#/components/schemas/Usuario"
          },
          "fecha": {
            "type": "string",
            "format": "date-time"
          },
          "detalles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DetalleVenta"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}